+++ docs/search/all_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['begin_7',['begin',['../class_sinric_pro_class.html#a57722dec8f1f987f57c8fff9eed66765',1,'SinricProClass']]],
+  ['begin_7',['begin',['../class_sinric_pro_class.html#ab19ffdad72d356ef0a32db068455a320',1,'SinricProClass']]],
   ['brightnesscallback_8',['BrightnessCallback',['../class_sinric_pro_light.html#a1bf03bd01ae1f4ddf164260435dc2026',1,'SinricProLight']]]
 ];