+++ docs/search/functions_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['begin_138',['begin',['../class_sinric_pro_class.html#a57722dec8f1f987f57c8fff9eed66765',1,'SinricProClass']]]
+  ['begin_138',['begin',['../class_sinric_pro_class.html#ab19ffdad72d356ef0a32db068455a320',1,'SinricProClass']]]
 ];