+++ src/SinricProTV.h
@@ -54,6 +54,7 @@ class SinricProTV :  public SinricProDevice {
      * @param[in]   deviceId    String which contains the ID of device
      * @param[in]   volumeDelta Integer with relative volume the device should change about (-100..100)
      * @param[out]  volumeDelta Integer with absolute volume device has been set to
+     * @param[in]   volumeDefault Bool `false` if the user specified the amount by which to change the volume; otherwise `true`
      * @return      the success of the request
      * @retval      true        request handled properly
      * @retval      false       request was not handled properly because of some error
@@ -61,7 +62,7 @@ class SinricProTV :  public SinricProDevice {
      * @section AdjustVolumeCallback_TV Example-Code
      * @snippet callbacks.cpp onAdjustVolume
      **/      
-    typedef std::function<bool(const String&, int&)> AdjustVolumeCallback;
+    typedef std::function<bool(const String&, int&, bool)> AdjustVolumeCallback;
 
     /**
      * @brief Callback definition for onMute function
@@ -215,7 +216,8 @@ bool SinricProTV::handleRequest(const DeviceId &deviceId, const char* action, Js
 
   if (adjustVolumeCallback && actionString == "adjustVolume") {
     int volume = request_value["volume"];
-    success = adjustVolumeCallback(deviceId, volume);
+    bool volumeDefault = request_value["volumeDefault"] | false;
+    success = adjustVolumeCallback(deviceId, volume, volumeDefault);
     response_value["volume"] = volume;
     return success;
   }